+++ build.gradle
@@ -44,10 +44,13 @@ dependencies {
 	runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.11.2'
 	runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.2'
 
-	// DB
+	// DB - h2
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'com.h2database:h2'
 	runtimeOnly 'com.h2database:h2'
+
+	// DB - redis
+	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
 	// websocket
 	implementation 'org.springframework.boot:spring-boot-starter-websocket'
