+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/adapters/output/redis/data/RefreshTokenRepository.java
@@ -1,0 +1,8 @@
+package com.yonggamsa.withsuyeonjung.user.framework.adapters.output.redis.data;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
+    RefreshToken findByUserId(String userId);
+}
