+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/redis/RedisConfig.java
@@ -1,0 +1,28 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.redis;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+
+@Configuration
+@EnableRedisRepositories
+public class RedisConfig {
+
+    private final String host = "localhost";
+    private final int port = 6379;
+
+    @Bean
+    public RedisConnectionFactory redisConnectionFactory(){
+        return new LettuceConnectionFactory(host, port);
+    }
+
+    @Bean
+    public RedisTemplate<?, ?> redisTemplate(){
+       RedisTemplate<?, ?> redisTemplate = new RedisTemplate<>();
+       redisTemplate.setConnectionFactory(redisConnectionFactory());
+       return redisTemplate;
+    }
+}
