+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/JwtConfig.java
@@ -1,0 +1,18 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth;
+
+import com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.provider.JwtProvider;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+
+    @Value("${jwt.secret}")
+    private String secret;
+
+    @Bean
+    public JwtProvider jwtProvider(){
+        return new JwtProvider(secret);
+    }
+}
