+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/exception/RestAuthenticationEntryPoint.java
@@ -1,0 +1,28 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.exception;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
+        authException.printStackTrace();
+
+        String url = request.getRequestURI();
+        log.info("[{}] 로 요청이 들어왔지만 인증 예외가 발생했습니다.", url);
+
+        log.info("비인가 에러에 의한 응답입니다. Message : {}", authException.getMessage());
+        response.sendError(
+                HttpServletResponse.SC_UNAUTHORIZED,
+                authException.getLocalizedMessage()
+        );
+    }
+}
