+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,0 +1,39 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.handler;
+
+import com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.repository.OAuth2AuthorizationRequestBasedOnCookieRepository;
+import com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.utils.CookieUtil;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class OAuth2AuthenticationFailureHandler extends SimpleUrlAuthenticationFailureHandler {
+
+    private final OAuth2AuthorizationRequestBasedOnCookieRepository authorizationRequestRepository;
+
+    @Override
+    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
+        String targetUrl = CookieUtil.getCookie(request, OAuth2AuthorizationRequestBasedOnCookieRepository.REDIRECT_URI_NAME_COOKIE_NAME)
+                .map(Cookie::getValue)
+                .orElse("/");
+
+        exception.printStackTrace();
+
+        targetUrl = UriComponentsBuilder.fromUriString(targetUrl)
+                .queryParam("error", exception.getLocalizedMessage())
+                .build().toUriString();
+
+        authorizationRequestRepository.removeAuthorizationRequest(request, response);
+
+        getRedirectStrategy().sendRedirect(request, response, targetUrl);
+    }
+}
