+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/handler/TokenAccessDeniedHandler.java
@@ -1,0 +1,25 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.handler;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.HandlerExceptionResolver;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class TokenAccessDeniedHandler implements AccessDeniedHandler {
+
+    private final HandlerExceptionResolver handlerExceptionResolver;
+
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+        //response.sendError(HttpServletResponse.SC_FORBIDDEN, accessDeniedException.getMessage());
+        handlerExceptionResolver.resolveException(request, response, null ,accessDeniedException);
+    }
+}
