+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/token/AuthToken.java
@@ -1,0 +1,89 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.token;
+
+import io.jsonwebtoken.*;
+import io.jsonwebtoken.security.SecurityException;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+import java.security.Key;
+import java.util.Date;
+
+
+@Slf4j
+@Getter
+@RequiredArgsConstructor
+public class AuthToken {
+
+    @Getter
+    private final Key key;
+    private final String token;
+    private static final String AUTHORITIES_KEY = "role";
+
+    public AuthToken(String id, Date expiry, Key key) {
+        this.key = key;
+        this.token = createAuthToken(id, expiry);
+    }
+
+    public AuthToken(String id, String role, Date expiry, Key key){
+        this.key = key;
+        this.token = createAuthToken(id, role, expiry);
+    }
+
+    private String createAuthToken(String id, Date expiry){
+        return Jwts.builder()
+                .setSubject(id)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .setExpiration(expiry)
+                .compact();
+    }
+
+    private String createAuthToken(String id, String role, Date expiry){
+        return Jwts.builder()
+                .setSubject(id)
+                .claim(AUTHORITIES_KEY, role)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .setExpiration(expiry)
+                .compact();
+    }
+
+    public boolean validate(){
+        return this.getTokenClaims() != null;
+    }
+
+    public Claims getTokenClaims(){
+        try{
+            return Jwts.parserBuilder()
+                    .setSigningKey(key)
+                    .build()
+                    .parseClaimsJws(token)
+                    .getBody();
+        } catch (SecurityException e){
+            log.info("Invalid JWT signature");
+        } catch (MalformedJwtException e){
+            log.info("Invalid JWT token");
+        } catch (ExpiredJwtException e){
+            log.info("Expired JWT token");
+        } catch (UnsupportedJwtException e){
+            log.info("Unsupported JWT token");
+        } catch (IllegalArgumentException e){
+            log.info("JWT token compact of handler are invalid");
+        }
+        return null;
+    }
+
+    public Claims getExpriedTokenClaims(){
+        try {
+            Jwts.parserBuilder()
+                    .setSigningKey(key)
+                    .build()
+                    .parseClaimsJws(token)
+                    .getBody();
+        } catch (ExpiredJwtException e){
+            log.info("Expired JWT token");
+            return e.getClaims();
+        }
+        return null;
+    }
+
+}
