+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/utils/CookieUtil.java
@@ -1,0 +1,62 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.utils;
+
+import org.springframework.util.SerializationUtils;
+
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Base64;
+import java.util.Optional;
+
+public class CookieUtil {
+
+    public static Optional<Cookie> getCookie(HttpServletRequest request, String name){
+        Cookie[] cookies = request.getCookies();
+
+        if(cookies != null && cookies.length > 0){
+            for(Cookie cookie : cookies){
+                if(name.equals(cookie.getName())){
+                    return Optional.of(cookie);
+                }
+            }
+        }
+        return Optional.empty();
+    }
+
+    public static void addCookie(HttpServletResponse response, String name, String value, int maxAge){
+        Cookie cookie = new Cookie(name, value);
+        cookie.setPath("/");
+        cookie.setHttpOnly(true);
+        cookie.setMaxAge(maxAge);
+
+        response.addCookie(cookie);
+    }
+
+    public static void deleteCookie(HttpServletRequest request, HttpServletResponse response, String name){
+        Cookie[] cookies = request.getCookies();
+
+        if(cookies != null && cookies.length > 0){
+            for(Cookie cookie : cookies){
+                if(name.equals(cookie.getName())){
+                    cookie.setValue("");
+                    cookie.setPath("/");
+                    cookie.setMaxAge(0);
+                    response.addCookie(cookie);
+                }
+            }
+        }
+    }
+
+    public static String serialize(Object obj){
+        return Base64.getUrlEncoder()
+                .encodeToString(SerializationUtils.serialize(obj));
+    }
+
+    public static <T> T deserialize(Cookie cookie, Class<T> clazz){
+        return clazz.cast(
+                SerializationUtils.deserialize(
+                        Base64.getUrlDecoder().decode(cookie.getValue()))
+        );
+    }
+
+}
