+++ src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/security/oauth/utils/HeaderUtil.java
@@ -1,0 +1,23 @@
+package com.yonggamsa.withsuyeonjung.user.framework.configuration.security.oauth.utils;
+
+import javax.servlet.http.HttpServletRequest;
+
+public class HeaderUtil {
+
+    private static final String HEADER_AUTHORIZATION = "Authorization";
+    private static final String TOKEN_PREFIX = "Bearer ";
+
+    public static String getAccessToken(HttpServletRequest request){
+        String headerValue = request.getHeader(HEADER_AUTHORIZATION);
+
+        if(headerValue == null) {
+            return null;
+        }
+
+        if(headerValue.startsWith(TOKEN_PREFIX)){
+            return headerValue.substring(TOKEN_PREFIX.length());
+        }
+
+        return null;
+    }
+}
