--- src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/spring/YamlPropertySourceFactory.java
@@ -1,24 +1,0 @@
-package com.yonggamsa.withsuyeonjung.user.framework.configuration.spring;
-
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
-import org.springframework.core.env.PropertiesPropertySource;
-import org.springframework.core.env.PropertySource;
-import org.springframework.core.io.support.EncodedResource;
-import org.springframework.core.io.support.PropertySourceFactory;
-
-import java.io.IOException;
-import java.util.Properties;
-
-public class YamlPropertySourceFactory implements PropertySourceFactory {
-
-    @Override
-    public PropertySource<?> createPropertySource(String name, EncodedResource resource) throws IOException {
-
-        YamlPropertiesFactoryBean factory = new YamlPropertiesFactoryBean();
-        factory.setResources(resource.getResource());
-
-        Properties properties = factory.getObject();
-
-        return new PropertiesPropertySource(resource.getResource().getFilename(), properties);
-    }
-}
