--- src/main/java/com/yonggamsa/withsuyeonjung/user/framework/configuration/spring/security/oauth/repository/OAuth2AuthorizationRequestBasedOnCookieRepository.java
@@ -1,56 +1,0 @@
-package com.yonggamsa.withsuyeonjung.user.framework.configuration.spring.security.oauth.repository;
-
-import com.nimbusds.oauth2.sdk.util.StringUtils;
-import com.yonggamsa.withsuyeonjung.user.framework.configuration.spring.security.oauth.utils.CookieUtil;
-import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
-import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class OAuth2AuthorizationRequestBasedOnCookieRepository implements AuthorizationRequestRepository<OAuth2AuthorizationRequest> {
-
-    public static final String OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME = "oauth2_auth_request";
-    public static final String REDIRECT_URI_NAME_COOKIE_NAME = "redirect_uri";
-    public static final String REFRESH_TOKEN = "refresh_token";
-    private static final int cookieExpireSeconds = 180;
-
-    @Override
-    public OAuth2AuthorizationRequest loadAuthorizationRequest(HttpServletRequest request) {
-        return CookieUtil.getCookie(request, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME)
-                .map(cookie -> CookieUtil.deserialize(cookie, OAuth2AuthorizationRequest.class))
-                .orElse(null);
-    }
-
-    @Override
-    public void saveAuthorizationRequest(OAuth2AuthorizationRequest authorizationRequest, HttpServletRequest request, HttpServletResponse response) {
-        if(authorizationRequest == null) {
-            CookieUtil.deleteCookie(request, response, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME);
-            CookieUtil.deleteCookie(request, response, REDIRECT_URI_NAME_COOKIE_NAME);
-            CookieUtil.deleteCookie(request, response, REFRESH_TOKEN);
-            return;
-        }
-
-        CookieUtil.addCookie(response, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME, CookieUtil.serialize(authorizationRequest), cookieExpireSeconds);
-        String redirectUriAfterLogin = request.getParameter(REDIRECT_URI_NAME_COOKIE_NAME);
-        if(StringUtils.isNotBlank(redirectUriAfterLogin)){
-            CookieUtil.addCookie(response, REDIRECT_URI_NAME_COOKIE_NAME, redirectUriAfterLogin, cookieExpireSeconds);
-        }
-    }
-
-    @Override
-    public OAuth2AuthorizationRequest removeAuthorizationRequest(HttpServletRequest request) {
-        return this.loadAuthorizationRequest(request);
-    }
-
-    @Override
-    public OAuth2AuthorizationRequest removeAuthorizationRequest(HttpServletRequest request, HttpServletResponse response) {
-        return this.loadAuthorizationRequest(request);
-    }
-
-    public void removeAuthorizationRequestCookies(HttpServletRequest request, HttpServletResponse response){
-        CookieUtil.deleteCookie(request, response, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME);
-        CookieUtil.deleteCookie(request, response, REDIRECT_URI_NAME_COOKIE_NAME);
-        CookieUtil.deleteCookie(request, response, REFRESH_TOKEN);
-    }
-}
